+++ lib/src/connectionhandling/server/mqtt_client_mqtt_server_connection_handler.dart
@@ -11,7 +11,8 @@ part of mqtt_server_client;
 ///  for server based connections.
 abstract class MqttServerConnectionHandler extends MqttConnectionHandlerBase {
   /// Initializes a new instance of the [MqttServerConnectionHandler] class.
-  MqttServerConnectionHandler();
+  MqttServerConnectionHandler({@required int maxConnectionAttempts})
+      : super(maxConnectionAttempts: maxConnectionAttempts);
 
   /// Use a websocket rather than TCP
   bool useWebSocket = false;